f-analytics - Converted to service only
