+++ packages/services/f-analytics/src/store/default-state.js
@@ -1,0 +1,34 @@
+export const platformDataDefault = {
+    environment: 'localhost',
+    name: undefined,
+    appType: undefined,
+    applicationId: undefined,
+    userAgent: undefined,
+    branding: undefined,
+    country: undefined,
+    language: undefined,
+    jeUserPercentage: undefined,
+    currency: undefined,
+    version: undefined,
+    instancePosition: undefined,
+    isPilot: undefined
+};
+
+export const userDataDefault = {
+    'a-UserId': undefined,
+    authType: undefined,
+    email: undefined,
+    globalUserId: undefined,
+    signinType: undefined,
+    signupDate: undefined
+};
+
+export const pageDataDefault = {
+    group: undefined,
+    name: undefined,
+    httpStatusCode: 200,
+    conversationId: undefined,
+    requestId: undefined,
+    orientation: undefined,
+    display: undefined
+};
